+++ module04/ex01/Brain.hpp
@@ -6,7 +6,7 @@
 /*   By: ozamora- <ozamora-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/13 15:53:52 by ozamora-          #+#    #+#             */
-/*   Updated: 2025/09/13 16:58:47 by ozamora-         ###   ########.fr       */
+/*   Updated: 2025/09/15 16:40:13 by ozamora-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,9 @@ class Brain
 		~Brain();
 		Brain& operator=(const Brain& src);
 
-		std::string getIdea(int index) const;
-		void setIdea(int index, const std::string& idea); // The input string is not a copy and cannot be modified 
-		void printIdea(int index);
+		std::string getIdea(size_t index) const;
+		void setIdea(size_t index, const std::string& idea); // The input string is not a copy and cannot be modified 
+		void printIdea(size_t index);
 		void printAllIdeas();
 };
